requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
nltk==3.8.1
spacy==3.6.1
streamlit==1.25.0
pdfplumber==0.9.0
python-docx==0.8.11
joblib==1.3.1
matplotlib==3.7.2
seaborn==0.12.2 